Remove server package
