+++ client-service/src/main/java/com/jaspersoft/android/sdk/service/ServerInfoService.java
@@ -1,4 +1,4 @@
-package com.jaspersoft.android.sdk.service.server;
+package com.jaspersoft.android.sdk.service;
 
 import android.support.annotation.VisibleForTesting;
