+++ client-service/src/main/java/com/jaspersoft/android/sdk/service/TokenProvider.java
@@ -22,38 +22,17 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.sdk.service.server;
+package com.jaspersoft.android.sdk.service;
 
 import android.support.annotation.NonNull;
-import android.support.annotation.VisibleForTesting;
-import android.support.annotation.WorkerThread;
 
-import com.jaspersoft.android.sdk.service.data.server.ServerInfo;
+import com.jaspersoft.android.sdk.network.api.auth.Token;
 
 /**
- * Always make call on server
- *
  * @author Tom Koptel
  * @since 2.0
  */
-final class GreedyInfoProvider implements InfoProvider {
-    private final ServerInfoService mServerInfoService;
-
-    @VisibleForTesting
-    GreedyInfoProvider(ServerInfoService serverInfoService) {
-        mServerInfoService = serverInfoService;
-    }
-
-    public static InfoProvider newInstance(String serverUrl) {
-        ServerInfoService service = ServerInfoService.newInstance(serverUrl);
-        return new GreedyInfoProvider(service);
-    }
-
-    @Override
+public interface TokenProvider {
     @NonNull
-    @WorkerThread
-    public ServerInfo provideInfo() {
-        return mServerInfoService.requestServerInfo()
-                .toBlocking().first();
-    }
+    Token<?> provideToken();
 }
