+++ client-service/src/main/java/com/jaspersoft/android/sdk/service/Utils.java
@@ -22,17 +22,20 @@
  * <http://www.gnu.org/licenses/lgpl>.
  */
 
-package com.jaspersoft.android.sdk.service.server;
-
-import android.support.annotation.NonNull;
-
-import com.jaspersoft.android.sdk.network.api.auth.Token;
+package com.jaspersoft.android.sdk.service;
 
 /**
  * @author Tom Koptel
  * @since 2.0
  */
-public interface TokenProvider {
-    @NonNull
-    Token<?> provideToken();
+final class Utils {
+    private Utils() {
+    }
+
+    public static <T> T checkNotNull(T object, String message) {
+        if (object == null) {
+            throw new NullPointerException(message);
+        }
+        return object;
+    }
 }
